+++ src/main/java/br/edu/fema/modelo/atividadesfixacao/domain/alimento/repository/TipoAlimentoRepository.java
@@ -3,6 +3,6 @@ package br.edu.fema.modelo.atividadesfixacao.domain.alimento.repository;
 import br.edu.fema.modelo.atividadesfixacao.domain.alimento.entity.TipoAlimento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoAlimentoRepository extends JpaRepository<TipoAlimento, Integer> {
+public interface TipoAlimentoRepository extends JpaRepository<TipoAlimento, Long> {
 
 }
